fix(vsdk-225): address demo UI review feedback
